+++ src/main/scala/dxWDL/DeclAttrs.scala
@@ -12,6 +12,15 @@ case class DeclAttrs(m: Map[String, JsValue]) {
             case _ => false
         }
     }
+
+    def getDefault: Option[JsValue] = {
+        m.get("default")
+    }
+
+    // add another attribute
+    def add(key:String, value:JsValue) : DeclAttrs = {
+        DeclAttrs(m + (key -> value))
+    }
 }
 
 object DeclAttrs {